pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "supabase>=2.15.0",
    "werkzeug>=3.1.3",
    "google-api-python-client>=2.166.0",
    "requests>=2.32.3",
    "python-telegram-bot>=22.0",
    "telegram>=0.0.1",
]
